Snapshot for pushing to github.
In its present state, ling is a web service that will sort and rank
words from JSON input text. In order to do so, it must first parse
a word frequency list from the American National Corpus using the
load-words function and save the data on disk in the form of Clojure
data structures. It expects to read a file containing these structures
from "words.clj" after which it parses any words it receives at the
API endpoints in core.clj and sorts them based on the ANC data for
any word matches found in the input text. Sorting is based on some
weird and probably wrong math I concocted.