Refactor to simplify routes, improve Python client.
Simplify route configuration and word API by pushing the language
processor (for parsing words and sentences) and the list of word
frequencies into module-level atoms in words.clj.

Fix the Python client and use it to write some basic functional tests
that hit all of the ling API endpoints with a test file of your choice
filled with words and/or sentences (e.g., a news article).